requirements.txt
appnope==0.1.2
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
click==7.1.2
decorator==4.4.2
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jupyter-client==6.1.11
jupyter-core==4.7.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.8.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.10
protobuf==3.14.0
ptyprocess==0.7.0
pycodestyle==2.6.0
Pygments==2.7.4
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.5
pyzmq==20.0.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
Flask==2.0.1
Flask-JWT-Extended==4.2.3
Flask-SQLAlchemy==3.0.0
psycopg2-binary==2.9.1
python-dotenv==0.19.1
